FISH-642 openmq-5.1.1.final.payara-p6
Workaround fix for OpenMQ breaking logging on Shoal/GMS clustered
instances. OpenMQ attempts to start a clustered broker but can run into an
error incurred from loopback addresses and attempt to fall back to a
non-clustered broker. This behaviour is flawed and destroys logging for
the instance, and so this workaround fix makes it so that OpenMQ simply
logs the error and exits.

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>