FISH-962 Don't use pre-generated config files.
Using pre-generated config files introduces issues with backwards and
forwards compatibility, since they may or may not be there, and they may
or may not have been changed between versions. We also don't upgrade the
config file, so this resolves that issue too.

So instead of using a pre-written one we generate them when using the
upgrade-server command, this way they will always be correct for a
release.

This commit also includes some other random fixes

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>