FISH-5955 Set app (system) class loader as the bundle parent CL
This allows OSGi bundles to access classes on the system class path.
However, bootdelegation rules apply (for both classes and resources)
and thus this effectively only allows access to classes in the oracle.sql
package if Oracle JDBC driver JAR is placed on the system classpath.
This is required for the EclipseLink Oracle extension to work.

The issue Issue #11129 is not longer present as Felix now mandates
bootdelegation rules also for resources and doesn't expose resources
in the parent classloader that aren't listed in the bootdelegation property.

(cherry picked from commit 7a2034a85e8e912c1534e5e350a98f70391ee126)